requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
narwhals==1.34.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
PyMuPDF==1.25.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
